Terminate individual tests after half an hour
While using the Automake parallel test harness, if a test hangs for
long enough for an external watchdog to kill the entire build process
(as happens in Debian sbuild after 150 minutes with no activity on
stdout/stderr), the logs will not be shown. If we make an individual
test time out sooner, logs are more likely to be shown.

Signed-off-by: Simon McVittie <smcv@debian.org>

Gbp-Pq: Name Terminate-individual-tests-after-half-an-hour.patch