Make corrupt-repo-ref.js executable
Debian's Lintian packaging consistency check complains that it isn't
executable but has a #! line. In fact it's reasonable to run this
script directly, so make it executable, and put it in a _scripts
variable so it will be installed executable.

Forwarded: https://github.com/ostreedev/ostree/pull/652
Applied-upstream: 2017.1

Gbp-Pq: Name Make-corrupt-repo-ref.js-executable.patch