Don't write to parent repo
In _try_clone_from_payload_link, don't try to do the clone in the
parent repo, because we don't want to modify that. parent repos are
typically used when you want a shared, immutable base.

For example in flatpak, the parent repo is the system repo which you
don't have write access to, so any modification to it will fail with
EACCES, making it impossible to install via the system helper.

Closes: #1524
Approved by: cgwalters
Applied-upstream: 2018.5, commit:28c7bc6d0e153a0b07bdb82d25473a490765067f

Gbp-Pq: Name Don-t-write-to-parent-repo.patch